deployment/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
Pillow>=9.0.0
matplotlib>=3.5.0
transformers>=4.30.0
tqdm>=4.64.0
scikit-learn>=1.0.0
pandas>=1.3.0 